+++ package.json
@@ -31,6 +31,7 @@
   },
   "dependencies": {
     "@vitejs/plugin-react": "^3.0.1",
+    "classnames": "^2.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "vite": "^4.0.4"
