+++ src/components/ColorTable/ColorTable.css
@@ -1,3 +1,17 @@
+#color-table {
+  grid-column: 1 / 3;
+  grid-row: 2;
+  display: grid;
+  grid-template-columns: fit-content(var(--max-width-main)) auto;
+  grid-template-rows: fit-content(100%) auto;
+}
+
+.color-comparison-header > div {
+  display: flex;
+  flex-direction: column;
+  width: var(--size-color-tile);
+}
+
 .lock-button-row-container > :last-child {
   justify-self: end;
 }
