--- src/components/DropDown.css
@@ -1,3 +1,0 @@
-.dropdown > label {
-  margin-right: var(--spacing-m);
-}
