+++ src/components/Lib/Button.tsx
@@ -3,6 +3,8 @@ import React from 'react';
 import Icon, { IconKey } from './Icon';
 import ToolTip from './ToolTip';
 
+import './Button.css';
+
 interface Props extends React.DetailedHTMLProps<React.ButtonHTMLAttributes<HTMLButtonElement>, HTMLButtonElement> {
   icon?: IconKey;
   tooltip?: string;
@@ -15,7 +17,9 @@ const Button: React.FC<Props> = ({ icon, tooltip, triggerProps, ...buttonProps }
       <button {...buttonProps}>{icon && <Icon icon={icon} />}</button>
     </ToolTip>
   ) : (
-    <button {...buttonProps}>{icon && <Icon icon={icon} />}</button>
+    <button className="focus-border" {...buttonProps}>
+      {icon && <Icon icon={icon} />}
+    </button>
   );
 
 export default Button;
