+++ src/components/Lib/Input.css
@@ -1,0 +1,21 @@
+input {
+  box-sizing: border-box;
+  text-align: center;
+  width: 100%;
+  background: 1px solid var(--color-bg);
+  border: var(--border-1px);
+  color: var(--color-text);
+  padding: 0.4rem 0.5rem;
+}
+
+/* Chrome, Safari, Edge, Opera */
+input::-webkit-outer-spin-button,
+input::-webkit-inner-spin-button {
+  -webkit-appearance: none;
+  margin: 0;
+}
+
+/* Firefox */
+input[type=number] {
+  -moz-appearance: textfield;
+}
